+++ .ci.yaml
@@ -3687,7 +3687,7 @@ targets:
 
   - name: Mac_ios flutter_gallery_ios__start_up_xcode_debug
     recipe: devicelab/devicelab_drone
-    presubmit: true
+    presubmit: false
     timeout: 60
     properties:
       tags: >
@@ -3697,6 +3697,7 @@ targets:
         {
           "sdk_version": "14c18"
         }
+    bringup: true
 
   - name: Mac_ios flutter_view_ios__start_up
     recipe: devicelab/devicelab_drone
@@ -3767,13 +3768,16 @@ targets:
 
   - name: Mac_ios integration_ui_ios_driver_xcode_debug
     recipe: devicelab/devicelab_drone
-    presubmit: false
+    presubmit: true
     timeout: 60
     properties:
       tags: >
         ["devicelab", "ios", "mac"]
       task_name: integration_ui_ios_driver_xcode_debug
-    bringup: true
+      $flutter/osx_sdk : >-
+        {
+          "sdk_version": "14c18"
+        }
 
   - name: Mac_ios integration_ui_ios_frame_number
     recipe: devicelab/devicelab_drone
